chore: tests ordering and tesseract removal
