+++ constants/__init__.py
@@ -2,3 +2,4 @@ from . import commands
 from .colors import *
 from .tesseract import *
 from .user import *
+from .social_links import *
