+++ ext/vp/communities/test_communities/test_create_community/mac/logo.png
