+++ ext/vp/communities/test_communities/test_create_community_channel/channel_icon_in_list.png
