+++ ext/vp/communities/test_communities/test_edit_community/button_updated_logo.png
