+++ ext/vp/communities/test_communities/test_edit_community/logo_in_settings_updated.png
