+++ ext/vp/communities/test_communities/test_edit_community/updated_logo.png
