+++ ext/vp/communities/test_communities/test_edit_community_channel/channel_icon_in_list.png
