+++ ext/vp/communities/test_communities/test_edit_community_channel/general_channel_icon_in_chat.png
