+++ ext/vp/onboarding/test_onboarding_generate_new_keys/test_generate_new_keys/jpeg_onboarding.png
