+++ ext/vp/onboarding/test_onboarding_generate_new_keys/test_generate_new_keys/png_profile.png
