+++ ext/vp/settings/settings_wallet/test_wallet_settings_account_order/test_change_account_order_by_drag_and_drop/watch_only_account_icon.png
