+++ gui/components/messaging/edit_group_name_and_image_popup.py
@@ -1,8 +1,8 @@
 import allure
 
 from gui.components.base_popup import BasePopup
-from gui.elements.qt.button import Button
-from gui.elements.qt.text_edit import TextEdit
+from gui.elements.button import Button
+from gui.elements.text_edit import TextEdit
 
 
 class EditGroupNameAndImagePopup(BasePopup):
