+++ gui/components/social_links_popup.py
@@ -15,7 +15,7 @@ class SocialLinksPopup(BasePopup):
         super(SocialLinksPopup, self).__init__()
         self._add_social_link_list_item = QObject('socialLink_StatusListItem')
         self._social_link_text_field = TextEdit('edit_TextEdit')
-        self._add_button = Button('add_StatusButton')
+        self._add_button = Button('o_StatusBackButton')
 
     @allure.step('Get social link')
     def _get_list_item(self, title: str) -> QObject:
