+++ tests/settings/settings_profile/test_settings_profile_change_password.py
@@ -1,0 +1,27 @@
+import allure
+import pytest
+from allure_commons._allure import step
+
+import constants
+from driver.aut import AUT
+from gui.main_window import MainWindow
+
+
+@allure.testcase('https://ethstatus.testrail.net/index.php?/cases/view/703005',
+                 'Change the password and login with new password')
+@pytest.mark.case(703005)
+@pytest.mark.parametrize('user_account, user_account_new',
+                         [pytest.param(constants.user.user_account_one, constants.user.user_account_one_changed_name)])
+def test_change_password_and_login(aut: AUT, main_screen: MainWindow, user_account, user_account_new):
+    with step('Open profile settings and change password'):
+        main_screen.left_panel.open_settings().left_panel.open_profile_settings().open_change_password_popup().change_password(
+            user_account.password, user_account_new.password)
+
+    with step('Restart application'):
+        aut.restart()
+        main_screen.authorize_user(user_account_new)
+
+    with step('Verify that the user logged in correctly'):
+        user_canvas = main_screen.left_panel.open_user_canvas()
+        profile_popup = user_canvas.open_profile_popup()
+        assert profile_popup.user_name == user_account_new.name
