+++ tests/wallet/test_wallet_main_manage_accounts.py
